Cloudkit-installer roll-up (#6)
* Replace commonLabels with updated syntax

When deploying these manifests, kustomize would produce the warning:

> Warning: 'commonLabels' is deprecated. Please use 'labels' instead. Run
> 'kustomize edit fix' to update your Kustomization automatically.

This commit replaces `commonLabels` with the updated `labels` syntax.

* Re-import operator from https://github.com/innabox/cloudkit-operator/

* Move template publisher to other aap related resources

* Remove ingresscontroller patch

* Remove explicit namespaces

We should only set the namespace on the top-level kustomization file.

* Rename fulfillment service service accounts

* Reorganize cloudkit-aap deployment

* Install AAP instance

Include the AnsibleAutomationPlatform manifest in base/cloudkit-aap.

* Manage images from top-level kustomization

* Make cloudikit.io/project a label

Previously, we were using a cloudkit.io/project annotation, but this does
not allow for label based filtering of resources.

* Replace fulfillment service manifests with upstream manifests

* Add fulfilment service route

Without the route there is no external access to the fulfillment service

* Reflect recent cloudkit-aap changes in bootstrap job

* Put sensible environment defaults in operator deployment

* Save typing, rename cloudkit-operator-controller-manager to cloudkit-controller

* Teach kustomize about certificate issuers

This instructs kustomize that spec/issueRef/name in a Certificate
corresponds to the name of an Issuer. This will allow issuer references to
get updated correctly when we rename resources using namePrefix/nameSuffix,
etc.

* Fix erroneous label

This was causing the controller pod to match the fulfillment-api service.

* Generate postgres password

Have the postgres pod generate a random password and put it in the
`fulfillemnt-database-password` Secret.

* Process environment variables in rules.yaml

Add a gomplate [1] init container to the fulfillment service that performs
environment variable substitution on rules.yaml

[1]: https://github.com/hairyhenderson/gomplate

* Set all images in base kustomization.yaml

* Remove old overlays

* Introduce new development overlay

* Rename all users of cloudkit.io to cloudkit.openshift.io

This is the namespace we were already using in the kubernetes operator; no
need to invent yet another one.